+++ tests/regression/test_cassettes_retain_global_configuration.py
@@ -1,7 +1,7 @@
-import os
+import pytest
 import unittest
 
-from betamax import Betamax
+from betamax import Betamax, cassette
 from requests import Session
 
 
@@ -11,9 +11,16 @@ class TestCassetteRecordMode(unittest.TestCase):
             config.default_cassette_options['record_mode'] = 'never'
 
     def tearDown(self):
-        os.unlink('tests/cassettes/regression_record_mode.json')
+        with Betamax.configure() as config:
+            config.default_cassette_options['record_mode'] = 'once'
 
     def test_record_mode_is_never(self):
         s = Session()
-        with Betamax(s).use_cassette('regression_record_mode') as recorder:
-            assert recorder.current_cassette.record_mode == 'never'
+        with pytest.raises(ValueError):
+            with Betamax(s) as recorder:
+                recorder.use_cassette('regression_record_mode')
+                assert recorder.current_cassette is None
+
+    def test_class_variables_retain_their_value(self):
+        opts = cassette.Cassette.default_cassette_options
+        assert opts['record_mode'] == 'never'